lost data structures and windows files, created new ones

+++ C/Assignment1/Calculator/add.c
@@ -1,5 +1,5 @@
 #include<stdio.h>
-double add(int a, int b)
+double add(float a, float b)
 {
 	return(a + b);
 }
