+++ C/Assignment1/Calculator/mul.c
@@ -1,5 +1,5 @@
 #include<stdio.h>
-double mul(int a, int b)
+double mul(float a, float b)
 {
 	return(a * b);
 }
